Update utilities.c, uint8_t not int in read/writeRegister

+++ src/utilities.c
@@ -13,7 +13,7 @@ void inc_PC (state *state){
     state -> currAddress += 0x4;
 }
 
-void writeRegister (int registerIndex, uint64_t newValue, uint8_t sf, uint64_t *generalRegisters) {
+void writeRegister (uint8_t registerIndex, uint64_t newValue, uint8_t sf, uint64_t *generalRegisters) {
     if (registerIndex == 31) {
         return;
     }
@@ -26,7 +26,7 @@ void writeRegister (int registerIndex, uint64_t newValue, uint8_t sf, uint64_t *
     }
 }
 
-uint64_t readRegister (int registerIndex, uint8_t sf, uint64_t *generalRegisters) {
+uint64_t readRegister (uint8_t registerIndex, uint8_t sf, uint64_t *generalRegisters) {
     if (registerIndex == 31) {
         return 0;
     }
